requirements.txt
discord-py~=2.6.4
uvicorn~=0.38.0
fastapi~=0.124.4
pydantic~=2.12.5
requests~=2.32.5